refactor: extract CLI commands from index.ts into commands/
